Resolve pylint warnings about cropcal_module.py.
Moves some functions from there into 3 new modules:
* ctsm/crop_calendars/check_constant_vars.py
* ctsm/crop_calendars/check_rx_obeyed.py
* ctsm/crop_calendars/convert_axis_time2gs.py
